+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Light bulbs".
 - Tests for "What the brainfuck !".
 - Tests for "Bouncing barry".
+- Tests for "Dice probability calculator".
 
 ## [1.11.0] - 2022-07-31
 ### Added